requirements.txt
click==8.2.1
Flask==3.1.1
Markdown==3.8.2
protobuf==6.31.1
python-dotenv==1.1.1
google.generativeai==0.8.5
google.auth==2.39.0
psycopg2==2.9.10